feat(aplicativo): monta fundação Kotlin Multiplatform dos módulos compartilhados
Recria os módulos Gradle Android como módulos KMP conforme a ADR-002:
:shared:core:{common,model,testing,designsystem}, :shared:domain:patrimonio,
:shared:app e :androidApp, com targets androidTarget, iosArm64 e
iosSimulatorArm64.

O diretório android/ passa a se chamar aplicativo/ (agora hospeda também o host
iOS) e o projeto Gradle raiz vira "savro". O design system migra para Compose
Multiplatform em commonMain, com fontes e strings servidos por compose.resources.
A tela de evidência sai da Activity e vira SavroApp() em :shared:app; a Activity
fica só como host.

verifyArchitecture e verifyDesignSystemTokens foram reescritas para o modelo de
source sets (commonMain/androidMain/iosMain), com regras distintas por source set
e a mesma allowlist de dependências adaptada aos novos módulos. A suíte TestKit
passa de 7 para 10 cenários.

Refs #193
