--- android/core/model/build.gradle.kts
@@ -1,3 +1,0 @@
-plugins {
-    alias(libs.plugins.kotlin.jvm)
-}
