--- android/domain/patrimonio/build.gradle.kts
@@ -1,8 +1,0 @@
-plugins {
-    alias(libs.plugins.kotlin.jvm)
-}
-
-dependencies {
-    implementation(project(":core:common"))
-    implementation(project(":core:model"))
-}
