--- android/settings.gradle.kts
@@ -1,23 +1,0 @@
-pluginManagement {
-    repositories {
-        google()
-        mavenCentral()
-        gradlePluginPortal()
-    }
-}
-
-dependencyResolutionManagement {
-    repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
-    repositories {
-        google()
-        mavenCentral()
-    }
-}
-
-rootProject.name = "savro-android"
-include(":app")
-include(":core:common")
-include(":core:model")
-include(":core:testing")
-include(":core:designsystem")
-include(":domain:patrimonio")
