+++ aplicativo/androidApp/src/main/AndroidManifest.xml
