+++ aplicativo/gradle.properties
