+++ aplicativo/shared/app/src/commonMain/composeResources/values/strings.xml
