+++ aplicativo/shared/core/designsystem/src/commonMain/composeResources/files/inter_ofl.txt
