+++ aplicativo/shared/core/designsystem/src/commonMain/composeResources/files/manrope_ofl.txt
