+++ aplicativo/shared/core/designsystem/src/commonMain/composeResources/font/inter_variable.ttf
