+++ aplicativo/shared/core/designsystem/src/commonMain/composeResources/font/manrope_variable.ttf
